pyproject.toml
[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
ignore = ["S607", "S101", "RET504"]
select = ["E", "F", "I", "UP", 'W', "B", "C", "T", "F"]

[tool.poetry]
line-length = 100
name = "my-rpi5-discord-bot"
version = "0.0.0"
description = ""
authors = ["c8763yee <rrrrr4413@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "23.2.1"
aiohttp = "3.9.3"
aiomqtt = "2.0.1"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
asyncio-atexit = "1.0.1"
attrs = "23.2.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
colorama = "0.4.6"
discord-py = "2.3.2"
distro = "1.9.0"
frozenlist = "1.4.1"
h11 = "0.14.0"
httpcore = "1.0.4"
httpx = "0.27.0"
idna = "3.6"
multidict = "6.0.5"
openai = "1.14.2"
paho-mqtt = "1.6.1"
psutil = "5.9.8"
pycparser = "2.21"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
pynacl = "1.5.0"
python-dotenv = "1.0.1"
requests = "2.31.0"
sniffio = "1.3.1"
tqdm = "4.66.2"
typing-extensions = "4.10.0"
urllib3 = "2.2.1"
yarl = "1.9.4"


[tool.poetry.group.dev.dependencies]
pylint = "3.1.0"
isort = "5.13.2"
black = "24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
